+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -7,6 +7,7 @@ unreleased
 +++++++++++++++++++
 * BC `vnet list-private-access-services`: renamed to `vnet list-endpoint-services`
 * BC `vnet subnet create/update`: renamed `--private-access-services` to `--service-endpoints`
+* `nsg rule create/update`: Add support for multiple IP ranges and port ranges.
 * `lb create`: Added support for SKU.
 * `public-ip create`: Added support for SKU.